refactor(projects): convert if-chain handler to declarative route table
Adopts @branch/lambda-http's dispatch() in place of the hand-rolled
if/endsWith routing. handler.ts is now a one-liner; routes.ts holds the
ordered Route[] table (bracketed by the ROUTES-START/END markers, now
here instead of handler.ts).

Split the 769-line handler into:
- controllers/projects.ts: list, get, create, update, delete
- controllers/dashboard.ts: dashboard, overview
- controllers/members.ts: members, assignable-staff
- controllers/donors.ts: project donors
- controllers/expenditures.ts: project expenditures
- services/projects.ts: loadProjectAggregates, syncMemberships,
  findUnknownUserIds, isProjectActive, indexByProject, projectIdFrom,
  toIsoDate (the former bottom-of-file helpers)

Route order preserves the original if-chain exactly: /projects/dashboard
and /projects/assignable-staff are declared before /projects/:id since
dispatch is first-match-wins and both are same-length path shapes that
:id would otherwise swallow. auth.ts, db.ts, validation-utils.ts,
swagger-utils.ts and dev-server.ts are untouched.

Each handler now authenticates via createAuthGuard(authenticateRequest)
at the AUTHENTICATED level, replicating the old global
"authContext.isAuthenticated && authContext.user" gate exactly (same
401 body). Domain authorization (canAccessProject/canEditProject/
canCreateProject/canDeleteProject/canListAssignableStaff) is unchanged
and still runs per-route from the untouched auth.ts.

No behaviour change: same status codes, messages, validation order,
SQL and transaction boundaries. The known dashboard anti-pattern (bulk
expenditure select + JS bucketing) is moved verbatim, not fixed.

Added a route-precedence regression test asserting GET /projects/dashboard
reaches the dashboard controller rather than GET /projects/:id.

Adds @branch/lambda-http as a dependency (file: to shared/lambda-http)
and regenerates package-lock.json.
